Refactoring headers in dem tests.
