+++ tests/core/simulation_control_04.cc
@@ -26,7 +26,7 @@
 #include <core/parameters.h>
 #include <core/simulation_control.h>
 
-// Tests
+// Tests (with common definitions)
 #include <../tests/tests.h>
 
 void
