+++ tests/core/solid_base_33.cc
@@ -18,8 +18,6 @@
 */
 
 // Deal.II includes
-#include <deal.II/base/utilities.h>
-
 #include <deal.II/distributed/tria.h>
 
 #include <deal.II/grid/grid_generator.h>
@@ -31,7 +29,7 @@
 #include <core/solid_base.h>
 #include <solvers/nitsche.h>
 
-// Tests
+// Tests (with common definitions)
 #include <../tests/tests.h>
 
 void
