+++ tests/dem/particle_particle_contact_force_linear.cc
@@ -17,11 +17,13 @@
  * Author: Shahab Golshan, Polytechnique Montreal, 2019-
  */
 
-// In this test, the performance of linear (Hookean) particle-particle contact
-// force  is checked
+/**
+ * @brief In this test, the performance of linear (Hookean) particle-particle
+ * contact force  is checked.
+ */
 
+// Deal.II
 #include <deal.II/base/parameter_handler.h>
-#include <deal.II/base/point.h>
 
 #include <deal.II/fe/mapping_q.h>
 
@@ -33,6 +35,7 @@
 #include <deal.II/particles/particle_handler.h>
 #include <deal.II/particles/particle_iterator.h>
 
+// Lethe
 #include <dem/dem_properties.h>
 #include <dem/dem_solver_parameters.h>
 #include <dem/find_cell_neighbors.h>
@@ -40,10 +43,8 @@
 #include <dem/pp_fine_search.h>
 #include <dem/pp_linear_force.h>
 
-#include <iostream>
-#include <vector>
-
-#include "../tests.h"
+// Tests (with common definitions)
+#include <../tests/tests.h>
 
 using namespace dealii;
