+++ tests/dem/particle_particle_fine_search.cc
@@ -17,11 +17,13 @@
  * Author: Shahab Golshan, Polytechnique Montreal, 2019-
  */
 
-// In this test, the performance of particle-particle fine search class is
-// evaluated
+/**
+ * @brief In this test, the performance of particle-particle fine search class
+ * is evaluated.
+ */
 
+// Deal.II
 #include <deal.II/base/parameter_handler.h>
-#include <deal.II/base/point.h>
 
 #include <deal.II/fe/mapping_q.h>
 
@@ -33,16 +35,14 @@
 #include <deal.II/particles/particle_handler.h>
 #include <deal.II/particles/particle_iterator.h>
 
-#include <dem/dem_solver_parameters.h>
+// Lethe
 #include <dem/find_cell_neighbors.h>
 #include <dem/pp_broad_search.h>
 #include <dem/pp_contact_info_struct.h>
 #include <dem/pp_fine_search.h>
 
-#include <iostream>
-#include <vector>
-
-#include "../tests.h"
+// Tests (with common definitions)
+#include <../tests/tests.h>
 
 using namespace dealii;
