+++ tests/dem/particle_wall_fine_search.cc
@@ -17,10 +17,14 @@
  * Author: Shahab Golshan, Polytechnique Montreal, 2019-
  */
 
-// In this test, the performance of particle-wall fine search is investigated
 
+/**
+ * @brief In this test, the performance of particle-wall fine search is
+ * investigated.
+ */
+
+// Deal.II
 #include <deal.II/base/parameter_handler.h>
-#include <deal.II/base/point.h>
 
 #include <deal.II/fe/mapping_q.h>
 
@@ -32,16 +36,14 @@
 #include <deal.II/particles/particle_handler.h>
 #include <deal.II/particles/particle_iterator.h>
 
-#include <dem/dem_solver_parameters.h>
+// Lethe
 #include <dem/find_boundary_cells_information.h>
 #include <dem/pw_broad_search.h>
 #include <dem/pw_contact_info_struct.h>
 #include <dem/pw_fine_search.h>
 
-#include <iostream>
-#include <vector>
-
-#include "../tests.h"
+// Tests (with common definitions)
+#include <../tests/tests.h>
 
 using namespace dealii;
